paint_app_ai/paint_agent/pyproject.toml
[project]
name = "paint-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.41.0",
    "mcp[cli]>=1.16.0",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.1.1",
]
